pyproject.toml
[project]
name = "literature-scout"
version = "0.1.0"
description = "Weekly skeletal muscle literature scouting and digest generation"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.0",
  "PyYAML>=6.0.0",
]

[tool.setuptools]
packages = ["literature_scout"]
